Add OrbitWorld, connect it to Agent
Now agents are Metaballs that have gravitational orbital mechanics
around a central metaball. Connected this to Agent, so the thrust
decision for each metaball is decided by the Agent, but the Agent
implementation is still just a Monte Carlo evaluator that always selects
a random action and treats the problem as a multi-arm bandit.